backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP client and async support
httpx==0.25.2
aiofiles==23.2.1

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AWS Integration
boto3==1.34.0
botocore==1.34.0

# Data processing and storage
orjson==3.9.10
tinydb==4.8.0

# Logging and monitoring
loguru==0.7.2

# Caching
cachetools==5.3.2

# Date and time utilities
python-dateutil==2.8.2

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Type checking (optional)
mypy==1.7.1